MX1 ADS: Rename files for better conformance to naming conventions
